move gs-vectortiles configuration to gs-spring-configuration
