+++ src/geoserver-configuration/src/main/java/org/geoserver/configuration/extension/vectortiles/VectorTilesGeoJsonConfiguration.java
@@ -3,18 +3,17 @@
  * application directory.
  */
 
-package org.geoserver.cloud.autoconfigure.extensions.vectortiles;
+package org.geoserver.configuration.extension.vectortiles;
 
 import org.geoserver.cloud.config.factory.ImportFilteredResource;
 import org.springframework.context.annotation.Configuration;
 
 /**
- * Configuration for TopoJSON Vector Tiles format.
+ * Configuration for GeoJSON Vector Tiles format.
  *
  * @since 2.27.0
  */
-@ConditionalOnVectorTilesTopoJson
-@ImportFilteredResource(
-        "jar:gs-vectortiles-.*!/applicationContext.xml#name=(wmsTopoJSONBuilderFactory|wmsTopoJSONMapOutputFormat)")
 @Configuration
-public class TopoJsonConfiguration {}
+@ImportFilteredResource(
+        "jar:gs-vectortiles-.*!/applicationContext.xml#name=(wmsGeoJsonBuilderFactory|wmsGeoJsonMapOutputFormat)")
+public class VectorTilesGeoJsonConfiguration {}
