+++ src/geoserver-configuration/src/main/java/org/geoserver/configuration/extension/vectortiles/VectorTilesMapBoxConfiguration.java
@@ -1,0 +1,19 @@
+/* (c) 2025 Open Source Geospatial Foundation - all rights reserved
+ * This code is licensed under the GPL 2.0 license, available at the root
+ * application directory.
+ */
+
+package org.geoserver.configuration.extension.vectortiles;
+
+import org.geoserver.cloud.config.factory.ImportFilteredResource;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Configuration for MapBox Vector Tiles format.
+ *
+ * @since 2.27.0
+ */
+@Configuration
+@ImportFilteredResource(
+        "jar:gs-vectortiles-.*!/applicationContext.xml#name=(wmsMapBoxBuilderFactory|wmsMapBoxMapOutputFormat)")
+public class VectorTilesMapBoxConfiguration {}
