+++ src/geoserver-configuration/src/main/java/org/geoserver/configuration/extension/vectortiles/VectorTilesTopoJsonConfiguration.java
@@ -3,18 +3,17 @@
  * application directory.
  */
 
-package org.geoserver.cloud.autoconfigure.extensions.vectortiles;
+package org.geoserver.configuration.extension.vectortiles;
 
 import org.geoserver.cloud.config.factory.ImportFilteredResource;
 import org.springframework.context.annotation.Configuration;
 
 /**
- * Configuration for MapBox Vector Tiles format.
+ * Configuration for TopoJSON Vector Tiles format.
  *
  * @since 2.27.0
  */
-@ConditionalOnVectorTilesMapBox
-@ImportFilteredResource(
-        "jar:gs-vectortiles-.*!/applicationContext.xml#name=(wmsMapBoxBuilderFactory|wmsMapBoxMapOutputFormat)")
 @Configuration
-public class MapBoxConfiguration {}
+@ImportFilteredResource(
+        "jar:gs-vectortiles-.*!/applicationContext.xml#name=(wmsTopoJSONBuilderFactory|wmsTopoJSONMapOutputFormat)")
+public class VectorTilesTopoJsonConfiguration {}
